requirements.txt
celery==5.2.7
celery-redbeat==2.1.0
cachetools==5.3.1
docker==6.1.3
requests==2.31.0